renamed root project title
