+++ README.md
@@ -1,1 +1,1 @@
-This Telegram bot's aim to manage business-processes of home-based bakery
+This App's aim to manage business-processes of home-based bakery
