--- bakery_bot/adapters/database/di.py
@@ -1,42 +1,0 @@
-from collections.abc import AsyncIterator
-
-from dishka import AnyOf, BaseScope, Component, Provider, Scope, provide
-from sqlalchemy.ext.asyncio import AsyncEngine, AsyncSession, async_sessionmaker
-
-from bakery_bot.adapters.database.storages.user import UserStorage
-from bakery_bot.adapters.database.uow import SqlalchemyUow
-from bakery_bot.adapters.database.utils import create_engine, create_sessionmaker
-from bakery_bot.domains.interfaces.storages.user import IUserStorage
-from bakery_bot.domains.uow import AbstractUow
-
-
-class DatabaseProvider(Provider):
-    def __init__(
-        self,
-        dsn: str,
-        debug: bool,
-        scope: BaseScope | None = None,
-        component: Component | None = None,
-    ) -> None:
-        self.dsn = dsn
-        self.debug = debug
-        super().__init__(scope=scope, component=component)
-
-    @provide(scope=Scope.APP)
-    async def engine(self) -> AsyncIterator[AsyncEngine]:
-        async with create_engine(dsn=self.dsn, debug=self.debug) as engine:
-            yield engine
-
-    @provide(scope=Scope.APP)
-    def session_factory(self, engine: AsyncEngine) -> async_sessionmaker[AsyncSession]:
-        return create_sessionmaker(engine=engine)
-
-    @provide(scope=Scope.REQUEST)
-    def uow(
-        self, session_factory: async_sessionmaker[AsyncSession]
-    ) -> AnyOf[SqlalchemyUow, AbstractUow]:
-        return SqlalchemyUow(session=session_factory())
-
-    @provide(scope=Scope.REQUEST)
-    def user_storage(self, uow: SqlalchemyUow) -> IUserStorage:
-        return UserStorage(session=uow.session)
