--- bakery_bot/adapters/database/storages/user.py
@@ -1,99 +1,0 @@
-from collections.abc import Sequence
-from datetime import UTC, datetime
-from typing import NoReturn
-from uuid import UUID
-
-from sqlalchemy import exists, func, insert, select, update
-from sqlalchemy.exc import DBAPIError, IntegrityError, NoResultFound
-from sqlalchemy.ext.asyncio import AsyncSession
-
-from bakery_bot.adapters.database.converters.user import convert_user_table_to_dto
-from bakery_bot.adapters.database.tables import UserTable
-from bakery_bot.application.exceptions import (
-    EntityAlreadyExistsException,
-    EntityNotFoundException,
-    StorageException,
-)
-from bakery_bot.domains.entities.user import (
-    CreateUser,
-    UpdateUser,
-    User,
-    UserListParams,
-)
-from bakery_bot.domains.interfaces.storages.user import IUserStorage
-
-
-class UserStorage(IUserStorage):
-    def __init__(self, session: AsyncSession) -> None:
-        self.__session = session
-
-    async def create(self, *, input_dto: CreateUser) -> User:
-        stmt = insert(UserTable).values(**input_dto.to_dict()).returning(UserTable)
-        try:
-            result = (await self.__session.scalars(stmt)).one()
-        except IntegrityError as e:
-            self.__raise_exception(e)
-        return convert_user_table_to_dto(result=result)
-
-    async def get_by_id(self, *, input_id: UUID) -> User | None:
-        stmt = select(UserTable).where(
-            UserTable.id == input_id, UserTable.deleted_at.is_(None)
-        )
-        result = await self.__session.scalar(stmt)
-        return convert_user_table_to_dto(result=result) if result else None
-
-    async def get_list(self, *, input_dto: UserListParams) -> Sequence[User]:
-        stmt = (
-            select(UserTable)
-            .where(UserTable.deleted_at.is_(None))
-            .limit(input_dto.limit)
-            .offset(input_dto.offset)
-        )
-        result = await self.__session.scalars(stmt)
-        return [convert_user_table_to_dto(result=r) for r in result]
-
-    async def count(self, *, input_dto: UserListParams) -> int:
-        stmt = (
-            select(func.count())
-            .select_from(UserTable)
-            .where(UserTable.deleted_at.is_(None))
-        )
-        return await self.__session.scalar(stmt) or 0
-
-    async def exists_by_id(self, *, input_id: UUID) -> bool:
-        stmt = select(
-            exists().where(UserTable.id == input_id, UserTable.deleted_at.is_(None))
-        )
-        return bool(await self.__session.scalar(stmt))
-
-    async def update_by_id(self, *, input_dto: UpdateUser) -> User:
-        stmt = (
-            update(UserTable)
-            .where(UserTable.id == input_dto.id)
-            .values(**input_dto.to_dict())
-            .returning(UserTable)
-        )
-        try:
-            result = (await self.__session.scalars(stmt)).one()
-        except NoResultFound as e:
-            raise EntityNotFoundException(entity=User, entity_id=input_dto.id) from e
-        except IntegrityError as e:
-            self.__raise_exception(e)
-        return convert_user_table_to_dto(result=result)
-
-    async def delete_by_id(self, *, input_id: UUID) -> None:
-        stmt = (
-            update(UserTable)
-            .where(UserTable.id == input_id)
-            .values(deleted_at=datetime.now(tz=UTC))
-        )
-        await self.__session.execute(stmt)
-
-    def __raise_exception(self, e: DBAPIError) -> NoReturn:
-        constraint = e.__cause__.__cause__.constraint_name  # type: ignore[union-attr]
-        match constraint:
-            case "uq__users__username":
-                raise EntityAlreadyExistsException("Username already exists") from e
-            case "uq__users__email":
-                raise EntityAlreadyExistsException("Email already exists") from e
-        raise StorageException(self.__class__.__name__) from e
