+++ docker-compose.yaml
@@ -1,11 +1,11 @@
 version: '3.8'
 
 services:
-  challenge_bot:
+  bakery:
     build:
       context: .
       dockerfile: Dockerfile
-    container_name: bakery_bot_app
+    container_name: bakery_app
     env_file:
       - .env
     ports:
