+++ pyproject.toml
@@ -1,5 +1,5 @@
 [tool.poetry]
-name = "bakery_bot"
+name = "bakery"
 version = "0.1.0"
 description = ""
 authors = [
@@ -19,20 +19,47 @@ asyncly = "^0.3.3"
 msgspec = "^0.18.6"
 SQLAlchemy = "^2.0.36"
 alembic = "^1.16.1"
-
+asyncpg = "^0.30.0"
+uvloop = "^0.21.0"
+granian = "^1.6.4"
+pydantic = {extras = ["email"], version = "^2.10.3"}
+aiobotocore = "^2.22.0"
 
 
 [tool.poetry.group.dev.dependencies]
-ruff = "^0.4.9"
-mypy = "^1.10.0"
-pre-commit = "^3.7.1"
-types-passlib = "^1.7.7.20240327"
-
+pytest = "^8.3.4"
+pytest-asyncio = "^0.24.0"
+pytest-cov = "^6.0.0"
+coverage = "^7.6.9"
+mypy = "^1.12.1"
+ruff = "^0.8.2"
+polyfactory = "^2.18.1"
+pre-commit = "^4.0.1"
+dirty-equals = "^0.8.0"
+httpx = "^0.28.1"
 
 [build-system]
 requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
 
+[tool.pytest.ini_options]
+asyncio_mode = "auto"
+python_files = "test_*"
+python_functions = "test_*"
+python_classes = "TestSuite*"
+addopts = "-p no:cacheprovider"
+asyncio_default_fixture_loop_scope = "function"
+
+[tool.coverage.run]
+branch = true
+source = ["bakery"]
+command_line = "-m pytest"
+
+[tool.coverage.report]
+show_missing = true
+
+[tool.coverage.xml]
+output = "coverage.xml"
 
 [tool.ruff]
 line-length = 88
@@ -73,7 +100,7 @@ ignore = ["ISC001"]
 fixable = ["ALL"]
 
 [tool.ruff.lint.isort]
-known-first-party = ["challenge_bot"]
+known-first-party = ["bakery", "tests"]
 
 [tool.mypy]
 check_untyped_defs = true
@@ -85,5 +112,6 @@ no_implicit_optional = true
 [[tool.mypy.overrides]]
 module = [
     "py_expression_eval.*",
+    "asyncpg.*",
 ]
 ignore_missing_imports = true
